Update imports for tests to pull the code straight from src
The __tests__ module used exports from the index.ts file to be able to
test if those exports would be actually available. Changed this to
improve agility in testing the code and pull directly from the piece
needed for testing. Also some imports were wrong from the changes made
on the previous commit so it needed a fix.

Relevance: 60
Tags: chore, fix